+++ lib/features/what_if/domain/entities/what_if_result.dart
@@ -1,8 +1,15 @@
+import 'package:decimal/decimal.dart';
 import 'package:equatable/equatable.dart';
 
+/// Bir günün varlık fiyatı.
+///
+/// `price` finansal alan olduğu için `Decimal`. IEEE-754 ondalık precision
+/// hatasıyla grafik tooltip'inde 1 kuruşluk fark görünmemesi için
+/// CLAUDE.md "para için double YASAK" kuralı burada da geçerli — fiyat
+/// tutarı tutar gibi davranır.
 class ChartPoint extends Equatable {
   final DateTime date;
-  final double price;
+  final Decimal price;
 
   const ChartPoint({required this.date, required this.price});
 
@@ -10,17 +17,27 @@ class ChartPoint extends Equatable {
   List<Object?> get props => [date, price];
 }
 
+/// Bir "ya alsaydım" hesaplamasının sonucu.
+///
+/// **Tip kuralı (CLAUDE.md "Yasak Listesi"):** Tüm para tutarları ve birim
+/// sayıları `Decimal`'dır. `double` IEEE-754 binary olduğu için ondalık
+/// toplama hatası verir (örn. `0.1 + 0.2 != 0.3`); kullanıcı 1 kuruşluk
+/// fark görse uygulama güvenilirliğini kaybeder.
+///
+/// **Yüzde alanları (`*Percent`)** `double` olarak kalır — sadece display
+/// için kullanılır, aggregasyon yok. `NumberFormat.decimalPercentPattern`
+/// zaten double bekler, gereksiz Decimal cast'i UI'i karmaşıklaştırır.
 class WhatIfResult extends Equatable {
   final String assetSymbol;
   final String assetDisplayName;
   final DateTime buyDate;
   final DateTime? sellDate;
-  final double buyPrice;
-  final double sellPrice;
-  final double unitsAcquired;
-  final double initialValueTry;
-  final double finalValueTry;
-  final double profitLossTry;
+  final Decimal buyPrice;
+  final Decimal sellPrice;
+  final Decimal unitsAcquired;
+  final Decimal initialValueTry;
+  final Decimal finalValueTry;
+  final Decimal profitLossTry;
   final double profitLossPercent;
   final bool isProfit;
   final List<ChartPoint> priceHistory;
