+++ lib/features/what_if/presentation/widgets/reverse_result_card.dart
@@ -222,15 +222,15 @@ class _ReverseResultCardState extends State<ReverseResultCard>
                 // ── Detaylar ──────────────────────────────────────────────
                 _Row(
                   l10n.resultBuyPrice,
-                  _tryFormatter.format(result.buyPrice),
+                  _tryFormatter.format(result.buyPrice.toDouble()),
                 ),
                 _Row(
                   l10n.resultSellPrice,
-                  _tryFormatter.format(result.sellPrice),
+                  _tryFormatter.format(result.sellPrice.toDouble()),
                 ),
                 _Row(
                   l10n.resultUnitsAcquired,
-                  _formatUnits(result.unitsAcquired),
+                  _formatUnits(result.unitsAcquired.toDouble()),
                 ),
                 _Row(l10n.resultDuration, _formatDuration(l10n)),
 
@@ -279,10 +279,12 @@ class _ReverseResultCardState extends State<ReverseResultCard>
                   ),
                   _AnimatedRow(
                     l10n.realProfitLoss,
-                    (result.requiredInvestmentTry *
-                            result.realProfitLossPercent! /
-                            100)
-                        .toDouble(),
+                    // Display-only çarpım; Decimal → double tek seferlik
+                    // dönüşüm `NumberFormat` zaten 2 ondalık haneye
+                    // yuvarladığı için precision farkı yaratmaz.
+                    result.requiredInvestmentTry.toDouble() *
+                        result.realProfitLossPercent! /
+                        100,
                     formatter: _trySignedFormatter,
                     valueColor: result.realProfitLossPercent! >= 0
                         ? AppColors.profit
